Split so that it can be used by albums app
